requirements.txt
absl-py==0.11.0
certifi==2020.6.20
chumpy==0.70
cycler==0.10.0
Cython==0.29.21
flatbuffers==1.12
future==0.18.2
jax==0.2.6
jaxlib==0.1.57
kiwisolver==1.3.1
matplotlib==3.3.3
numpy==1.19.4
opencv-python==4.4.0.46
opendr==0.78
opt-einsum==3.3.0
Pillow==8.0.1
pygame==2.0.0
pyparsing==2.4.7
pyrealsense2==2.40.0.2483
python-dateutil==2.8.1
scipy==1.5.4
six==1.15.0
torch==1.5.0
torchvision==0.6.0
transforms3d==0.3.1
